Move generate_csr code out of generate_cert.
