+++ security/pkg/pki/ca/generate_cert.go
@@ -77,23 +77,6 @@ type CertOptions struct {
 // URIScheme is the URI scheme for Istio identities.
 const URIScheme = "spiffe"
 
-// GenCSR generates a X.509 certificate sign request and private key with the given options.
-func GenCSR(options CertOptions) ([]byte, []byte, error) {
-	// Generates a CSR
-	priv, err := rsa.GenerateKey(rand.Reader, options.RSAKeySize)
-	if err != nil {
-		return nil, nil, fmt.Errorf("CSR generation fails at RSA key generation (%v)", err)
-	}
-	template := GenCSRTemplate(options)
-	csrBytes, err := x509.CreateCertificateRequest(rand.Reader, &template, crypto.PrivateKey(priv))
-	if err != nil {
-		return nil, nil, fmt.Errorf("CSR generation fails at X509 cert request generation (%v)", err)
-	}
-
-	csr, privKey := encodePem(true, csrBytes, priv)
-	return csr, privKey, nil
-}
-
 func fatalf(template string, args ...interface{}) {
 	log.Errorf(template, args)
 	os.Exit(-1)
@@ -123,18 +106,6 @@ func GenCert(options CertOptions) ([]byte, []byte) {
 	return encodePem(false, certBytes, priv)
 }
 
-func encodePem(isCSR bool, csrOrCert []byte, priv *rsa.PrivateKey) ([]byte, []byte) {
-	encodeMsg := "CERTIFICATE"
-	if isCSR {
-		encodeMsg = "CERTIFICATE REQUEST"
-	}
-	csrOrCertPem := pem.EncodeToMemory(&pem.Block{Type: encodeMsg, Bytes: csrOrCert})
-
-	privDer := x509.MarshalPKCS1PrivateKey(priv)
-	privPem := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: privDer})
-	return csrOrCertPem, privPem
-}
-
 // LoadSignerCredsFromFiles loads the signer cert&key from the given files.
 //   signerCertFile: cert file name
 //   signerPrivFile: private key file name
@@ -161,6 +132,18 @@ func LoadSignerCredsFromFiles(signerCertFile string, signerPrivFile string) (*x5
 	return cert, key, nil
 }
 
+func encodePem(isCSR bool, csrOrCert []byte, priv *rsa.PrivateKey) ([]byte, []byte) {
+	encodeMsg := "CERTIFICATE"
+	if isCSR {
+		encodeMsg = "CERTIFICATE REQUEST"
+	}
+	csrOrCertPem := pem.EncodeToMemory(&pem.Block{Type: encodeMsg, Bytes: csrOrCert})
+
+	privDer := x509.MarshalPKCS1PrivateKey(priv)
+	privPem := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: privDer})
+	return csrOrCertPem, privPem
+}
+
 func genSerialNum() *big.Int {
 	serialNumLimit := new(big.Int).Lsh(big.NewInt(1), 128)
 	serialNum, err := rand.Int(rand.Reader, serialNumLimit)
@@ -168,22 +151,6 @@ func genSerialNum() *big.Int {
 		fatalf("Serial number generation failure (%v)", err)
 	}
 	return serialNum
-}
-
-// GenCSRTemplate generates a certificateRequest template with the given options.
-func GenCSRTemplate(options CertOptions) x509.CertificateRequest {
-	template := x509.CertificateRequest{
-		Subject: pkix.Name{
-			Organization: []string{options.Org},
-		},
-	}
-
-	if h := options.Host; len(h) > 0 {
-		s := buildSubjectAltNameExtension(h)
-		template.ExtraExtensions = []pkix.Extension{*s}
-	}
-
-	return template
 }
 
 // genCertTemplate generates a certificate template with the given options.
