+++ security/pkg/pki/ca/generate_cert_test.go
@@ -16,8 +16,6 @@ package ca
 
 import (
 	"crypto/x509"
-	"encoding/pem"
-	"strings"
 	"testing"
 	"time"
 
@@ -26,54 +24,6 @@ import (
 )
 
 var now = time.Now().Round(time.Second).UTC()
-
-func TestGenCSR(t *testing.T) {
-	// Options to generate a CSR.
-	csrOptions := CertOptions{
-		Host:       "test_ca.com",
-		Org:        "MyOrg",
-		RSAKeySize: 512,
-	}
-
-	csrPem, _, err := GenCSR(csrOptions)
-
-	if err != nil {
-		t.Errorf("failed to gen CSR")
-	}
-
-	pemBlock, _ := pem.Decode(csrPem)
-	if pemBlock == nil {
-		t.Errorf("failed to decode csr")
-	}
-	csr, err := x509.ParseCertificateRequest(pemBlock.Bytes)
-	if err != nil {
-		t.Errorf("failed to parse csr")
-	}
-	if err = csr.CheckSignature(); err != nil {
-		t.Errorf("csr signature is invalid")
-	}
-	if csr.Subject.Organization[0] != "MyOrg" {
-		t.Errorf("csr subject does not match")
-	}
-	if !strings.HasSuffix(string(csr.Extensions[0].Value[:]), "test_ca.com") {
-		t.Errorf("csr host does not match")
-	}
-}
-
-func TestGenCSRWithInvalidOption(t *testing.T) {
-	// Options with invalid Key size.
-	csrOptions := CertOptions{
-		Host:       "test_ca.com",
-		Org:        "MyOrg",
-		RSAKeySize: -1,
-	}
-
-	csr, priv, err := GenCSR(csrOptions)
-
-	if err == nil || csr != nil || priv != nil {
-		t.Errorf("Should have failed")
-	}
-}
 
 func TestLoadSignerCredsFromFiles(t *testing.T) {
 	testCases := map[string]struct {
